refactor(tables): export all table related components and interfaces from index file
